+++ plugins/quick-brick-xray/README.md
@@ -4,19 +4,81 @@ This package contains the code to use XRay logger in a react-native app.
 
 ## How to use it
 
-Simply import the module and use the available methods to report events
+Simply import the logger and create an instance of the logger by providing the category and subsystem. With that instance, you can send events with different log levels.
 
 ```typescript
 import XRayLogger from "@applicaster/quick-brick-xray";
 
-XRayLogger.log({
-  category: "my cool app",
-  subsystem: "app/core/init",
+const logger = new XRayLogger("my app", "app");
+
+logger.log({
   message: "app launched !",
 });
 ```
 
-If the native module isn't available, logs will be directed to the console instead. Logs are also sent to the console as well when the app is running with `__DEV__` flag on
+If the native module isn't available, logs will be directed to the console instead. Logs are also sent to the console as well when the app is running with React Native's `__DEV__` flag on.
+
+Available methods for logging are `log`, `debug`, `info`, `warn` and `error`. The argument for each logging method is an object with the following properties:
+
+```typescript
+type XRayEventData = {
+  message: string;
+  data?: AnyDictionary;
+  error?: Error;
+};
+```
+
+You can also assign context to a logger. When logging an event, the context data will be logged along with the other properties of the event. Additional context passed will be merged with the existing context. If your logger has a parent, the parent context will be included as well.
+
+```typescript
+logger.addContext({ foo: "bar" });
+```
+
+You can extend loggers and add different subsystems. This can be useful to handle different contexts
+
+```typescript
+// creates a logger with `app/a` subsystem
+const aLogger = logger.addSubsystem("a");
+aLogger.addContext({ someProp: "foo" });
+aLogger.log({ message: "some log" });
+
+// will have { foo: "bar", someProp: "foo" }; in context
+
+// creates another logger with `app/b` subsystem
+const bLogger = logger.addSubsystem("b");
+bLogger.addContext({ bar: "baz" });
+bLogger.log({ message: "some log" });
+
+// will have { foo: "bar", bar: "baz" }; in context
+```
+
+Loggers also allow you to create events which will be sent at a later point, allowing to gather and add data in different steps and places
+
+```typescript
+// creates an empty event with verbose log level
+const event = aLogger.createEvent();
+
+// you can then assign all the properties you want
+
+event
+  .setLogLevel(XRayLogLevel.info)
+  .setMessage("some log")
+  .addData({ propToLog: "some value" });
+
+// data added is also merged with the existing data
+
+event.addData({ otherProp: "other value" });
+
+// you can also attach an error object, mainly if using error level
+
+event
+  .setLogLevel(XRayLogLevel.error)
+  .attachError(new Error("An error occurred"));
+
+// when the event is ready to be sent, it can be logged
+
+event.send();
+```
 
 ## API
 
@@ -41,21 +103,30 @@ type XRayEvent = {
   exception?: Error;
 };
 
-type XRayLogEvent<LogLevel> = XRayEvent & {
-  level: LogLevel;
-};
+declare interface XRayEventI {
+  logger: XRayLoggerI;
+  level: XRayLogLevel;
+  message: string;
+  data: AnyDictionary;
+  error?: Error;
+  setLevel(level: XRayLogLevel): this;
+  setMessage(message: string): this;
+  addData(data: AnyDictionary): this;
+  attachError(error: Error): this;
+  send(): void;
+}
 
-type XRayVerboseEvent = XRayLogEvent<XRayLogLevel.verbose>;
-type XRayDebugEvent = XRayLogEvent<XRayLogLevel.debug>;
-type XRayInfoEvent = XRayLogEvent<XRayLogLevel.info>;
-type XRayWarningEvent = XRayLogEvent<XRayLogLevel.warning>;
-type XRayErrorEvent = XRayLogEvent<XRayLogLevel.error>;
-
-interface XRayLoggerI {
-  log: (event: XRayEvent) => void;
-  debug: (event: XRayEvent) => void;
-  info: (event: XRayEvent) => void;
-  warning: (event: XRayEvent) => void;
-  error: (event: XRayEvent) => void;
+declare interface XRayLoggerI {
+  category: string;
+  subsystem: string;
+  context: AnyDictionary;
+  log(event: XRayEvent): void;
+  debug(event: XRayEvent): void;
+  info(event: XRayEvent): void;
+  warning(event: XRayEvent): void;
+  error(event: XRayEvent): void;
+  addContext(context: AnyDictionary): this;
+  addSubsystem(subsystem: string): XRayLoggerI;
+  createEvent(): XRayEventI;
 }
 ```
